requirements.txt
Django==1.8.4
Pillow==2.9.0
South==1.0.2
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
django-paypal==0.2.5
django-rest-framework-mongoengine==2.0.2
djangorestframework==3.2.3
futures==3.0.3
mongoengine==0.9.0
pymongo==2.8.1
pysqlite==2.8.1
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
websocket-client==0.32.0
wsgiref==0.1.2
flup
boto
gcs-oauth2-boto-plugin
gsutil
